refactor: move default image data to another file
